+++ CHANGES.md
@@ -24,6 +24,7 @@
 - Replace Worker's `all_keys` with `get_workers_keys` on `RQ` main class
 - Replace Workers's `count` with `get_workers_count` on `RQ` main class
 - RENAME: `registry.py` -> `registries.py`
+- RENAME: `command.py` -> `commands.py`
 
 ### RQ 1.15.1 (2023-06-20)